\name{tgthesis-package}
\alias{tgthesis-package}
\alias{tgthesis}
\docType{package}
\title{
Data and Functions to reproduce my thesis
}
\description{
Data and Functions to reproduce my thesis
}
\details{
\tabular{ll}{
Package: \tab tgthesis\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-08-05\cr
License: \tab GPL-3\cr
}
}
\author{
Tobias Gauster

Maintainer: Tobias Gauster <t.gauster@koop-wasser.at>
}
\references{
Literature or other references for background information
}
\keyword{ package }
